Move histogram to avoidance pkg
