+++ avoidance/CMakeLists.txt
@@ -44,6 +44,7 @@ include_directories(
 
 ## Declare a C++ library
 set(AVOIDANCE_CPP_FILES   "src/common.cpp"
+                          "src/histogram.cpp"
 )
 if(NOT DISABLE_SIMULATION)
   set(AVOIDANCE_CPP_FILES "${AVOIDANCE_CPP_FILES}"
